Run Miri tests in CI and rustfmt check
